print binary

+++ bit_manipulation/1-print_binary.c
@@ -17,7 +17,7 @@ void print_binary(unsigned long int n)
 		return;
 	}
 
-	for (i = sizeof(unsigned int) * 8 - 1; i >= 0; i--)
+	for (i = sizeof(unsigned long int) * 8 - 1; i >= 0; i--)
 	{
 		if (n & (1UL << i))
 		{
